refactor(cli): correct declaration homing and dedupe flag wiring (review)
Addresses review feedback on the cluster.go split:

- Re-home declarations the split had placed in the previous command's file so
  each lives with its command: deleteLongDesc, diagnoseLongDesc, listLongDesc,
  startLongDesc, diffLongDesc, the restore errors, and the switch
  errors/history. The create-only k3s flag constants and newCreateLifecycleConfig
  move from connect.go to create.go.
- Dedupe the --name/--provider flag wiring shared by info, diagnose and the
  lifecycle simple commands into lifecycle.BindNameAndProviderFlags, clearing the
  jscpd clone the split surfaced (verified: 0 clones).
- Remove a stale NOTE comment in connect.go and a detached doc comment above
  handleCreateRunE in create.go.
- restore: only print "Restored <type>" when every file for that type restored;
  partial failures are already surfaced per-file and via ErrRestoreFailed.
